test(config): consolidate iterative tests into the table-driven file
config_test.go was already table-driven, just under a different file
name. Its coverage is a subset of what was config_table_test.go,
with three small gaps closed here:

* TestUpdaterConfigDefaults now asserts LocalTrustedRoot equals the
  rootBytes passed to New, and RemoteTargetsURL is derived as
  RemoteMetadataURL + "/targets". Both were checked in the old
  TestNewUpdaterConfig via EqualExportedValues.

* TestEnsurePathsExistTable gains an errorIs field, and the empty-
  paths case now asserts the returned error is errors.Is os.ErrNotExist
  (the stricter check the old TestEnsurePathsExist made).

The old config_test.go is removed and config_table_test.go is renamed
back to config_test.go now that there is only one test file per the
project convention.

go test -cover ./metadata/config/... stays at 28.6% of statements.
